Moved example music app into it's own Swift files.
