+++ FunctionalViewControllers/Supporting-Files/AppDelegate.swift
@@ -10,34 +10,6 @@ import UIKit
 
 // TODO: a ViewController should be only generic over the result!
 
-struct Album {
-    let name: String
-}
-
-struct Artist {
-    let name : String
-    let additionalInformation : String
-    let albums: [Album]
-}
-
-let artists : [Artist] = [
-    Artist(name: "JS Bach", additionalInformation: "Some more info", albums: [Album(name: "The Art of Fugue")]),
-    Artist(name: "Simeon Ten Holt", additionalInformation: "Bla bla", albums: [])
-]
-
-let chooseArtist: ViewController<[Artist], Artist> = tableViewController { cell, artist in
-    cell.textLabel!.text = artist.name
-    return cell
-}
-
-let chooseAlbum: ViewController<[Album],Album> = tableViewController { cell, album in
-    cell.textLabel?.text = album.name
-    return cell
-}
-
-let navigation = rootViewController(chooseArtist).map { $0.albums } >>> chooseAlbum
-
-
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
